Refactor CSAT-related imports and functions for better organization and security
